--- views/testyelp.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-
-	<title>Test Yelp</title>
-	<meta charset="utf-8">
-
-	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" integrity="sha384-1q8mTJOASx8j1Au+a5WDVnPi2lkFfwwEAa8hDDdjZlpLegxhjVME1fgjWPGmkzs7" crossorigin="anonymous">
-	<script src="https://code.jquery.com/jquery.js"></script>
-	<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js" integrity="sha384-0mSbJDEHialfmuBBQP6A4Qrprq5OVfW37PRR3j5ELqxss1yVqOtnepnHVP9aJ7xS" crossorigin="anonymous"></script>
-
-</head>
-
-<body>
-
-	<div class = row>
-
-		<div class = col-md-3>
-
-			<h2>Name</h2>
-			
-		</div>
-
-		<div class = col-md-3>
-
-			<h2>Location</h2>
-			
-		</div>
-
-		<div class = col-md-3>
-
-			<h2>Open</h2>
-			
-		</div>
-
-		<div class = col-md-3>
-
-			<h2>Rating</h2>
-			
-		</div>
-		
-	</div><!-- close row -->
-
-</body>
-</html>
